fixed the query mapping
Former-commit-id: 2e191cefab0fa2922e75e38a5fa8c11b3666cdee

+++ fda-metadata-db/entities/src/main/java/at/tuwien/entities/database/query/Query.java
@@ -46,7 +46,7 @@ public class Query {
     private String resultHash;
 
     @Column
-    private Integer resultNumber;
+    private Long resultNumber;
 
 
 }
